LayoutTests/imported/w3c/web-platform-tests/clipboard-apis/async-write-image-read-image.https-expected.txt

Body needed for test_driver.click()
The bottom image should display the same image as the top image.

Original Image:


Image after copy/paste:



FAIL Verify write and read clipboard [image/png Blob] assert_equals: expected 1 but got 0
FAIL Verify write error on malformed data [image/png ClipboardItem] promise_rejects_dom: function "function() { throw e }" threw object "NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission." that is not a DOMException DataError: property "name" is equal to "NotAllowedError", expected "DataError"
